// src/app/model/response/teacher-profile-response.ts
export interface TeacherProfileResponse {
  teacherUserId: number;
  addressId: number;
  teacherUsername: string;
  firstName: string;
  lastname: string;
  phoneNumber: string;
  email: string;
  gender: string;
  dateOfBirth: string;
  dateOfJoin: string;
  subjectName: string;
  doorNum: string;
  street: string;
  addrLine: string;
  city: string;
  state: string;
  pincode: number;
}